FIX Make sure that kernel lives entirely in the top-left matrix half
This fixes a bug I head when visualizing the kernel's mass conservation
properties. Here, the plot yielded very different results:
1. Plot without first symmetrizing the kernel layers.
2. Symmetrize with `K = [.5*(K_k + K_k.T) for K_k in K]`.

Case 2 lead to the correct results, Case 1 did not.

This was due to the fact that my kernel definition was incorrect, and
only fixed by symmetrizing. The `K_loss` term should not be a straight
line crossing the diagonal. Instead, it should "bend" there.

This is to make sure that kernel lives entirely in the top-left matrix half.