FEAT Add test asserting equivalence of sampled & unsampled kernel
Let `N_m` be the number of bins on the discretized mass axis.
Then, the kernel has `N_m^3` entries in total, encoding the results
of `N_m^2` collisions.

These collisions, represented as index pairs $(i,j)$, are sampled.
Here, let `N_s` be the number of sampled collisions.

In the limit of `N_s -> N_m^2`, the sampled kernel should exactly
equal the unsampled one (assuming each point can be sampled at most once).

This seems to be the case, at least in the case of `N = const.`, where `N`
is the number of particles per mass bin, and is used to calculated the
weights, probability, & thus the outcome of the sampling.